refactor: 과목 관련 훅 위치 변경
