+++ src/pages/CourseSearchActivity/components/CourseSearchResult.tsx
@@ -3,8 +3,8 @@ import { useSuspenseQuery } from '@tanstack/react-query';
 import { getSearchedCourses } from '@/api/courses';
 import { Mixpanel } from '@/bootstrap/mixpanel';
 import { SelectableCourseItem } from '@/components/CourseItem/SelectableCourseItem';
+import { useCombinedCourses } from '@/hooks/course/useCombinedCourses';
 import { useAlertDialog } from '@/hooks/useAlertDialog';
-import { useCombinedCourses } from '@/hooks/useCombinedCourses';
 import { useToast } from '@/hooks/useToast';
 import { CourseSelectionChangeActionPayload } from '@/pages/CourseSearchActivity/type';
 import { CourseType } from '@/schemas/courseSchema';
